Get the events from a meetup group

Get the events from a meetup group

Usage

get_events(
  urlname,
  status = NULL,
  date_before = NULL,
  date_after = NULL,
  max_results = NULL,
  handle_multiples = "list",
  extra_graphql = NULL,
  ...
)

Arguments

urlname

Character. The name of the group as indicated in the https://www.meetup.com/ url.

status

Character vector of event statuses to retrieve.

date_before

Datetime string in "YYYY-MM-DDTHH:MM:SSZ" (ISO8601) format. Events occurring before this date/time will be returned.

date_after

Datetime string in "YYYY-MM-DDTHH:MM:SSZ" (ISO8601) format. Events occurring after this date/time will be returned.

max_results

Maximum number of results to return. If set to NULL, will return all available results (may take a long time).

handle_multiples

Character. How to handle multiple matches. One of "list" or "first", or "error". If "list", return a list-column with all matches. If "first", return only the first match.

extra_graphql

A graphql object. Extra objects to return

...

Should be empty. Used for parameter expansion

Value

A tibble with the events for the specified group

Examples

get_events("rladies-lagos", "past")
#> Warning: Selecting ‘env’ backend. Secrets are stored in environment variables
#> Warning: Selecting ‘env’ backend. Secrets are stored in environment variables
#> # A tibble: 13 × 23
#>    id    title event_url created_time        status date_time           duration
#>    <chr> <chr> <chr>     <dttm>              <chr>  <dttm>              <chr>   
#>  1 2641… satR… https://… 2019-08-20 11:54:50 PAST   2019-09-21 08:00:00 PT7H    
#>  2 2667… Data… https://… 2019-11-26 17:00:29 PAST   2019-12-07 08:00:00 PT7H    
#>  3 2678… Gett… https://… 2020-01-10 12:56:39 PAST   2020-02-22 08:00:00 PT7H    
#>  4 2843… Data… https://… 2022-03-01 00:27:32 PAST   2022-03-26 13:00:00 PT2H    
#>  5 2846… R La… https://… 2022-03-14 16:37:46 PAST   2022-05-07 15:00:00 PT1H    
#>  6 2922… Wome… https://… 2023-03-15 19:29:08 PAST   2023-03-24 09:00:00 PT2H    
#>  7 2931… R Pa… https://… 2023-04-27 23:25:21 PAST   2023-05-06 08:00:00 PT2H    
#>  8 2947… Unlo… https://… 2023-07-11 23:50:20 PAST   2023-07-22 15:30:00 PT2H    
#>  9 2950… Hack… https://… 2023-07-24 18:24:33 PAST   2023-08-14 16:00:00 PT30H   
#> 10 2939… Gene… https://… 2023-06-01 14:20:57 PAST   2023-09-09 14:00:00 PT2H    
#> 11 3000… Buil… https://… 2024-03-26 12:37:08 PAST   2024-04-06 16:00:00 PT3H    
#> 12 3003… Inte… https://… 2024-04-11 19:14:08 PAST   2024-04-25 10:00:00 PT5H    
#> 13 3000… Inte… https://… 2024-03-26 13:08:58 PAST   2024-04-27 16:00:00 PT3H    
#> # ℹ 16 more variables: description <chr>, group_id <chr>, group_name <chr>,
#> #   group_urlname <chr>, venues_id <list>, venues_name <list>,
#> #   venues_address <list>, venues_city <list>, venues_state <list>,
#> #   venues_postal_code <list>, venues_country <list>, venues_lat <list>,
#> #   venues_lon <list>, venues_venue_type <list>, rsvps_count <int>,
#> #   featured_event_photo_url <chr>
get_events(
   "rladies-lagos",
   status = "past",
   date_before = "2023-01-01T12:00:00Z"
)
#> Warning: Selecting ‘env’ backend. Secrets are stored in environment variables
#> Warning: Selecting ‘env’ backend. Secrets are stored in environment variables
#> # A tibble: 5 × 23
#>   id     title event_url created_time        status date_time           duration
#>   <chr>  <chr> <chr>     <dttm>              <chr>  <dttm>              <chr>   
#> 1 26415… satR… https://… 2019-08-20 11:54:50 PAST   2019-09-21 08:00:00 PT7H    
#> 2 26675… Data… https://… 2019-11-26 17:00:29 PAST   2019-12-07 08:00:00 PT7H    
#> 3 26780… Gett… https://… 2020-01-10 12:56:39 PAST   2020-02-22 08:00:00 PT7H    
#> 4 28432… Data… https://… 2022-03-01 00:27:32 PAST   2022-03-26 13:00:00 PT2H    
#> 5 28461… R La… https://… 2022-03-14 16:37:46 PAST   2022-05-07 15:00:00 PT1H    
#> # ℹ 16 more variables: description <chr>, group_id <chr>, group_name <chr>,
#> #   group_urlname <chr>, venues_id <chr>, venues_name <chr>,
#> #   venues_address <chr>, venues_city <chr>, venues_state <chr>,
#> #   venues_postal_code <chr>, venues_country <chr>, venues_lat <dbl>,
#> #   venues_lon <dbl>, venues_venue_type <chr>, rsvps_count <int>,
#> #   featured_event_photo_url <chr>