Get the current meetup members from a meetup group

Get the current meetup members from a meetup group

get_members(urlname, ..., extra_graphql = NULL, token = meetup_token())

Arguments

urlname

Required urlname of the Meetup group

...

Should be empty. Used for parameter expansion

extra_graphql

A graphql object. Extra objects to return

token

Meetup token

Value

A tibble with the following columns:

id

name

member_url

photo_link

status

role

created

most_recent_visit

References

https://www.meetup.com/api/schema/#GroupMembership https://www.meetup.com/api/schema/#User

Examples

if (FALSE) {
members <- get_members("rladies-remote")
}