Authorize meetupr

Authorize meetupr via the OAuth API. You will be directed to a web browser, asked to sign in to your Meetup account, and to grant meetupr permission to operate on your behalf. By default, these user credentials are cached in a file named .httr-oauth in the current working directory, from where they can be automatically refreshed, as necessary.

meetup_auth(
  token = NULL,
  new_user = FALSE,
  key = getOption("meetupr.consumer_key"),
  secret = getOption("meetupr.consumer_secret"),
  cache = getOption("meetupr.httr_oauth_cache"),
  verbose = TRUE
)

Arguments

token
optional; an actual token object or the path to a valid token stored as an .rds file

new_user
logical, defaults to FALSE. Set to TRUE if you want to wipe the slate clean and re-authenticate with the same or different Google account. This disables the .httr-oauth file in current working directory.

key, secret
the "Client ID" and "Client secret" for the application; defaults to the ID and secret built into the googlesheets package

cache
logical indicating if googlesheets should cache credentials in the default cache file .httr-oauth

verbose
logical; do you want informative messages?

Details

Most users, most of the time, do not need to call this function explicitly -- it will be triggered by the first action that requires authorization. Even when called, the default arguments will often suffice. However, when necessary, this function allows the user to

TODO: force the creation of a new token

TODO: retrieve current token as an object, for possible storage to an .rds file

TODO: read the token from an object or from an .rds file

TODO: provide your own app key and secret -- this requires setting up a new project in Google Developers Console

TODO: prevent caching of credentials in .httr-oauth

In a direct call to meetup_auth, the user can provide the token, app key and secret explicitly and can dictate whether interactively-obtained credentials will be cached in .httr_oauth. If unspecified, these arguments are controlled via options, which, if undefined at the time meetupr is loaded, are defined like so:

key

Set to option meetupr.consumer_key, which defaults to a consumer key that ships with the package

secret

Set to option meetupr.consumer_secret, which defaults to a consumer secret that ships with the package

cache

Set to option meetupr.httr_oauth_cache, which defaults to TRUE

To override these defaults in persistent way, predefine one or more of them with lines like this in a .Rprofile file:

options(meetupr.consumer_key = "FOO",
        meetupr.consumer_secret = "BAR",
        meetupr.httr_oauth_cache = FALSE)

See Startup for possible locations for this file and the implications thereof.

More detail is available from Authenticating with the Meetup API.

See also

Examples

if (FALSE) {
## load/refresh existing credentials, if available
## otherwise, go to browser for authentication and authorization
gs_auth()

## store token in an object and then to file
ttt <- gs_auth()
saveRDS(ttt, "ttt.rds")

## load a pre-existing token
gs_auth(token = ttt)       # from an object
gs_auth(token = "ttt.rds") # from .rds file
}