meetupr: Access Meetup Data

Provides programmatic access to the 'Meetup' GraphQL API (https://www.meetup.com/api/schema/), enabling users to retrieve information about groups, events, and members from 'Meetup' (https://www.meetup.com/). Supports authentication via 'OAuth2' and includes functions for common queries and data manipulation tasks.

See also

Author

Maintainer: Athanasia Mo Mowinckel a.m.mowinckel@psykologi.uio.no (ORCID)

Authors:

Erin LeDell oss@ledell.org

Olga Mierzwa-Sulima olga@rladies.org

Lucy D'Agostino McGowan lucy@rladies.org

Claudia Vitolo

Other contributors:

Gabriela De Queiroz gabriela@rladies.org [contributor]

Michael Beigelmacher [contributor]

Augustina Ragwitz augustina.ragwitz@ibm.com [contributor]

Greg Sutcliffe github@emeraldverie.org [contributor]

Rick Pack rickeyhp@gmail.com [contributor]

Ben Ubah ubah.ben22@gmail.com [contributor]

Maëlle Salmon maelle.salmon@yahoo.se (ORCID) [contributor]

Barret Schloerke barret@rstudio.com (ORCID) [contributor]

R-Ladies Global [copyright holder]