Get the events from a meetup group

Get the events from a meetup group

get_events(
  urlname,
  event_status = c("upcoming", "cancelled", "draft", "past", "proposed", "suggested"),
  fields = NULL,
  verbose = getOption("meetupr.verbose", rlang::is_interactive())
)

Arguments

urlname

Character. The name of the group as indicated in the https://meetup.com url.

event_status

Character or character vector (e.g. "upcoming" or c("past", "upcoming")). Event type - defaults to "upcoming". Valid inputs are:

cancelled

draft

past

proposed

suggested

upcoming

fields

Character, character vector or characters separated by comma (e.g "event_hosts" or c("event_hosts","attendance_count") or "event_hosts, group_past_event_count").

verbose

logical; do you want informative messages? TRUE by default in interactive sessions. Can be toggled by the meetupr.verbose option.

Value

A tibble with the following columns:

id

name

created

status

time

local_date

local_time

waitlist_count

yes_rsvp_count

venue_id

venue_name

venue_lat

venue_lon

venue_address_1

venue_city

venue_state

venue_zip

venue_country

description

link

resource

References

https://www.meetup.com/meetup_api/docs/:urlname/events/#list

Examples

if (FALSE) {
urlname <- "rladies-nashville"
past_events <- get_events(urlname = urlname,
                      event_status = "past")
upcoming_events <- get_events(urlname = urlname,
                      event_status = "upcoming")
past_meetings <- get_events(urlname = urlname,
                 event_status = "past",
                 fields = "event_hosts")
# get events hosts (co-organizers) of single past meeting
single_event <- past_meetings$resource[[1]]$event_hosts

# get all event hosts names (2) and host_counts (6) for that single event
# host_counts represents how events the person has co-organized or hosted.
do.call("rbind", lapply(single_event, '[', c(2,6)))
}