Authorize meetupr

Authorize meetupr via the OAuth API. You will be directed to a web browser, asked to sign in to your Meetup account, and to grant meetupr permission to operate on your behalf. By default, these user credentials are saved to a file in your home directory whose path is saved in .Renviron as MEETUPR_PAT. If you set set_env to FALSE but cache to TRUE, they are cached in a file named .httr-oauth in the current working directory. To control where the file is saved, use token_path.

If token is not already available, call meetup_auth to either load from cache or initiate OAuth2.0 flow. Return the token -- not "bare" but, rather, prepared for inclusion in downstream requests.

Suspend meetupr's authorization to place requests to the Meetup APIs on behalf of the authenticated user.

meetup_auth(
  token = meetup_token_path(),
  new_user = FALSE,
  key = getOption("meetupr.consumer_key"),
  secret = getOption("meetupr.consumer_secret"),
  cache = getOption("meetupr.httr_oauth_cache"),
  verbose = TRUE,
  set_renv = getOption("meetupr.set_renv"),
  token_path = NULL
)

meetup_token(verbose = FALSE)

meetup_deauth(clear_cache = TRUE, verbose = TRUE)

meetup_token_path()

Arguments

token
optional; an actual token object or the path to a valid token stored as an .rds file

new_user
logical, defaults to FALSE. Set to TRUE if you want to wipe the slate clean and re-authenticate with the same or different Meetup account. This disables the .httr-oauth file in current working directory.

key, secret
the "Client ID" and "Client secret" for the application; defaults to the ID and secret built into the meetupr package

cache
logical indicating if meetupr should cache credentials in the default cache file .httr-oauth or token_path.

verbose
logical; do you want informative messages?

set_renv
Logical indicating whether to save the created token as the default environment meetup token variable. Defaults to TRUE, meaning the token is saved to user's home directory as either the user provided path, or ".meetup_token.rds" (or, if that already exists, then .meetup_token1.rds or .meetup_token2.rds, etc.) and the path to the token to said token is then set in the user's .Renviron file and re- read to start being used in current active session. If cache is FALSE this is ignored.

token_path
Path where to save the token. If set_renv is FALSE, this is ignored.

clear_cache
logical indicating whether to disable the .httr-oauth file in working directory, if such exists, by renaming to .httr-oauth-SUSPENDED

Value

a request object (an S3 class provided by httr)

Either NULL or the path in which the token is saved.

Examples

if (FALSE) {
## load/refresh existing credentials, if available
## otherwise, go to browser for authentication and authorization
meetup_auth()

## store token in an object and then to file
ttt <- meetup_auth()
saveRDS(ttt, "ttt.rds")

## load a pre-existing token
meetup_auth(token = ttt)       # from an object
meetup_auth(token = "ttt.rds") # from .rds file
}
if (FALSE) {
meetup_token()
}
if (FALSE) {
meetup_deauth()
}
meetup_token_path()

#> NULL