Meetup API Authentication

Functions to manage authentication with the Meetup API. Includes functions to authenticate, and deauthorize by removing cached credentials.

Usage

meetup_auth(...)

meetup_deauth(
  client_name = Sys.getenv("MEETUP_CLIENT_NAME", "meetupr"),
  clear_keyring = TRUE
)

Arguments

...

Additional arguments to meetup_client().

client_name

A string representing the name of the client. By default, it is set to "meetupr" and retrieved from the MEETUP_CLIENT_NAME environment variable.

clear_keyring

A logical value indicating whether to clear the associated keyring entries. Defaults to TRUE.

Value

Nothing. Outputs messages indicating the result of the process.

Functions

meetup_auth(): Authenticate and display the authenticated user's name.

meetup_deauth(): Remove cached authentication for the Meetup API client.

Examples

if (FALSE) { # \dontrun{
meetup_auth()

# Default deauthorization
meetup_deauth()

# Deauthorization with a custom client name
meetup_deauth(client_name = "custom_client")
} # }